Store length with each memo entry. Don't use max_position
Max position determines the region of the buffer that must remain stable
for the result to stay in the memo table, but this can include area
beyond the area explicitly covered by the match (because of
backtracking). It can't be relied upon to compute the length of the
match, so an explicit length is stored instead.